Improvements
1) Track My Order :
- Need ""x"" close button next to page title (like we do on product detail page, etc.)...this would redirect user back to previous page
- Estimated Delivery date should show as ""Monday, July 16"" not ""Jul 16, Monday""
- When coming from News & Alerts page (by clicking on ""your item(s) have shipped"" alert notification, Estimated Delivery field is showing error".

+++ src/consumer-app/app/components/myorders/myorders.component.ts
@@ -298,7 +298,7 @@ export class MyordersComponent implements OnInit {
 
   getPurchaseDate(date) {
     let objDate = new Date(date), locale = "en-us", month = objDate.toLocaleString(locale, { month: "long" });
-    return objDate.toLocaleString(locale, { month: "short" }) + ' ' + objDate.getDate() + ', ' + this.formatAMPM(objDate);
+    return objDate.toLocaleString(locale, { month: "long" }) + ' ' + objDate.getDate() + ', ' + this.formatAMPM(objDate);
   }
 
   formatAMPM(date) {
